Added C export for DECX_core_CPU

+++ srcs/modules/CV/filters/CPU/bilateral_filter_exec.cxx
@@ -32,13 +32,13 @@ namespace decx{
             */
             static _THREAD_CALL_
                 decx::conv::_v256_2f32 _bilateral_uint8_f32_loop_in_kernel(const double* src, const float* _exp_chart_dist, const float* _exp_chart_diff,
-                    const uint2 ker_dims, const ushort reg_WL, const size_t Wsrc, const uint _loop);
+                    const uint2 ker_dims, const ushort reg_WL, const uint64_t Wsrc, const uint32_t _loop);
 
 
 
             static _THREAD_CALL_ decx::conv::_v256_2f32
                 _bilateral_uchar4_f32_loop_in_kernel(const float* src, const float* _exp_chart_dist, const float* _exp_chart_diff,
-                    const uint2 ker_dims, const ushort reg_WL, const size_t Wsrc, const uint _loop);
+                    const uint2 ker_dims, const ushort reg_WL, const uint64_t Wsrc, const uint32_t _loop);
         }
     }
 }
@@ -52,22 +52,22 @@ namespace decx
         namespace CPUK
         {
             static _THREAD_CALL_ void _bilateral_rect_fixed_uint8_ST(const double* src, const float* _exp_vals_dist, const float* _exp_chart_diff, double* dst,
-                const uint2 ker_dims, const uint Wsrc, const uint Wdst, const ushort reg_WL, const uint _loop);
+                const uint2 ker_dims, const uint32_t Wsrc, const uint32_t Wdst, const ushort reg_WL, const uint32_t _loop);
 
 
 
             static _THREAD_CALL_ void _bilateral_rect_flex_uint8_ST(const double* src, const float* _exp_vals_dist, const float* _exp_chart_diff, double* dst,
-                const uint2 proc_dim, const uint2 ker_dims, const uint Wsrc, const uint Wdst, const ushort reg_WL, const uint _loop);
+                const uint2 proc_dim, const uint2 ker_dims, const uint32_t Wsrc, const uint32_t Wdst, const ushort reg_WL, const uint32_t _loop);
 
 
 
             static _THREAD_CALL_ void _bilateral_rect_fixed_uchar4_ST(const float* src, const float* _exp_vals_dist, const float* _exp_chart_diff, float* dst,
-                const uint2 ker_dims, const uint Wsrc, const uint Wdst, const ushort reg_WL, const uint _loop);
+                const uint2 ker_dims, const uint32_t Wsrc, const uint32_t Wdst, const ushort reg_WL, const uint32_t _loop);
 
 
 
             static _THREAD_CALL_ void _bilateral_rect_flex_uchar4_ST(const float* src, const float* _exp_vals_dist, const float* _exp_chart_diff, float* dst,
-                const uint2 proc_dim, const uint2 ker_dims, const uint Wsrc, const uint Wdst, const ushort reg_WL, const uint _loop);
+                const uint2 proc_dim, const uint2 ker_dims, const uint32_t Wsrc, const uint32_t Wdst, const ushort reg_WL, const uint32_t _loop);
         }
     }
 }
@@ -80,8 +80,8 @@ decx::vis::CPUK::_bilateral_uint8_f32_loop_in_kernel(const double* __restrict   
                                                       const float* __restrict       _exp_chart_diff,
                                                       const uint2                   ker_dims, 
                                                       const ushort                  reg_WL, 
-                                                      const size_t                  Wsrc,
-                                                      const uint                    _loop)
+                                                      const uint64_t                  Wsrc,
+                                                      const uint32_t                    _loop)
 {
     uint8_t _store_reg[32];
     __m128i _proc_reg;
@@ -101,10 +101,10 @@ decx::vis::CPUK::_bilateral_uint8_f32_loop_in_kernel(const double* __restrict   
 
     __m256 _W_reg;
 
-    for (uint i = 0; i < ker_dims.y; ++i) 
+    for (uint32_t i = 0; i < ker_dims.y; ++i) 
     {
         _Y = 0;
-        for (uint j = 0; j < _loop; ++j) 
+        for (uint32_t j = 0; j < _loop; ++j) 
         {
             _mm256_store_pd((double*)_store_reg, _mm256_loadu_pd(src + i * Wsrc + j * 2));
 #ifdef _MSC_VER
@@ -150,8 +150,8 @@ decx::vis::CPUK::_bilateral_uchar4_f32_loop_in_kernel(const float* __restrict   
                                                       const float* __restrict       _exp_chart_diff,
                                                       const uint2                   ker_dims, 
                                                       const ushort                  reg_WL, 
-                                                      const size_t                  Wsrc,
-                                                      const uint                    _loop)
+                                                      const uint64_t                  Wsrc,
+                                                      const uint32_t                    _loop)
 {
     __m128i _proc_reg, aux_reg;
     
@@ -173,7 +173,7 @@ decx::vis::CPUK::_bilateral_uchar4_f32_loop_in_kernel(const float* __restrict   
     for (int i = 0; i < ker_dims.y; ++i) 
     {
         _Y = 0;
-        for (uint j = 0; j < _loop; ++j) 
+        for (uint32_t j = 0; j < _loop; ++j) 
         {
             if (j == 0) {
                 _proc_reg = _mm_castps_si128(_mm_load_ps(src + i * Wsrc));
@@ -230,13 +230,13 @@ decx::vis::CPUK::_bilateral_rect_fixed_uint8_ST(const double* __restrict   src,
                                         const float* __restrict       _exp_chart_diff,
                                         double* __restrict               dst,
                                         const uint2                     ker_dims,
-                                        const uint                      Wsrc,
-                                        const uint                      Wdst,
+                                        const uint32_t                      Wsrc,
+                                        const uint32_t                      Wdst,
                                         const ushort                    reg_WL,
-                                        const uint                      _loop)
+                                        const uint32_t                      _loop)
 {
     decx::conv::_v256_2f32 res_vec8;
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     __m256i _iv1, _iv2;
 
@@ -271,13 +271,13 @@ decx::vis::CPUK::_bilateral_rect_flex_uint8_ST(const double* __restrict     src,
                                                double* __restrict           dst,
                                                const uint2                  proc_dim,
                                                const uint2                  ker_dims,
-                                               const uint                   Wsrc,
-                                               const uint                   Wdst,
+                                               const uint32_t                   Wsrc,
+                                               const uint32_t                   Wdst,
                                                const ushort                 reg_WL,
-                                               const uint                   _loop)
+                                               const uint32_t                   _loop)
 {
     decx::conv::_v256_2f32 res_vec8;
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     __m256i _iv1, _iv2;
 
@@ -310,13 +310,13 @@ decx::vis::CPUK::_bilateral_rect_fixed_uchar4_ST(const float* __restrict   src,
                                         const float* __restrict       _exp_chart_diff,
                                         float* __restrict               dst,
                                         const uint2                     ker_dims,
-                                        const uint                      Wsrc,
-                                        const uint                      Wdst,
+                                        const uint32_t                      Wsrc,
+                                        const uint32_t                      Wdst,
                                         const ushort                    reg_WL,
-                                        const uint                      _loop)
+                                        const uint32_t                      _loop)
 {
     decx::conv::_v256_2f32 res_vec8;
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     __m256i _iv1, _iv2;
 
@@ -353,13 +353,13 @@ decx::vis::CPUK::_bilateral_rect_flex_uchar4_ST(const float* __restrict     src,
                                                float* __restrict           dst,
                                                const uint2                  proc_dim,
                                                const uint2                  ker_dims,
-                                               const uint                   Wsrc,
-                                               const uint                   Wdst,
+                                               const uint32_t                   Wsrc,
+                                               const uint32_t                   Wdst,
                                                const ushort                 reg_WL,
-                                               const uint                   _loop)
+                                               const uint32_t                   _loop)
 {
     decx::conv::_v256_2f32 res_vec8;
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     __m256i _iv1, _iv2;
 
@@ -392,23 +392,23 @@ void decx::vis::_bilateral_uint8_ST(const double* __restrict     src,
                              double* __restrict     dst, 
                              const uint2           proc_dim, 
                              const uint2           ker_dims,
-                             const uint            Wsrc,
-                             const uint            Wdst,
+                             const uint32_t            Wsrc,
+                             const uint32_t            Wdst,
                              const ushort          reg_WL,
-                             const uint            _loop)
+                             const uint32_t            _loop)
 {
     __m256 res_vec8;
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     decx::utils::frag_manager f_mgrH, f_mgrW;
     decx::utils::frag_manager_gen_from_fragLen(&f_mgrH, proc_dim.y, _BLOCKED_CONV2_UINT8_H_);
     decx::utils::frag_manager_gen_from_fragLen(&f_mgrW, proc_dim.x, _BLOCKED_CONV2_UINT8_W_);
 
-    const uint _loopH = f_mgrH.is_left ? f_mgrH.frag_num - 1 : f_mgrH.frag_num;
+    const uint32_t _loopH = f_mgrH.is_left ? f_mgrH.frag_num - 1 : f_mgrH.frag_num;
     
     for (int i = 0; i < _loopH; ++i) 
     {
-        const uint _loopW = f_mgrW.is_left ? f_mgrW.frag_num - 1 : f_mgrW.frag_num;
+        const uint32_t _loopW = f_mgrW.is_left ? f_mgrW.frag_num - 1 : f_mgrW.frag_num;
 
         for (int k = 0; k < _loopW; ++k) {
             decx::vis::CPUK::_bilateral_rect_fixed_uint8_ST(
@@ -418,7 +418,7 @@ void decx::vis::_bilateral_uint8_ST(const double* __restrict     src,
                 ker_dims, Wsrc, Wdst, reg_WL, _loop);
         }
         if (f_mgrW.is_left) {
-            const uint _sum_prev_lenW = proc_dim.x - f_mgrW.frag_left_over;
+            const uint32_t _sum_prev_lenW = proc_dim.x - f_mgrW.frag_left_over;
             decx::vis::CPUK::_bilateral_rect_flex_uint8_ST(
                 DECX_PTR_SHF_XY_SAME_TYPE(src, i * _BLOCKED_CONV2_UINT8_H_, _sum_prev_lenW * 2, Wsrc * 2),
                 _exp_chart_dist, _exp_chart_diff,
@@ -430,8 +430,8 @@ void decx::vis::_bilateral_uint8_ST(const double* __restrict     src,
     
     if (f_mgrH.is_left)
     {
-        const uint _sum_prev_lenH = proc_dim.y - f_mgrH.frag_left_over;
-        const uint _loopW = f_mgrW.is_left ? f_mgrW.frag_num - 1 : f_mgrW.frag_num;
+        const uint32_t _sum_prev_lenH = proc_dim.y - f_mgrH.frag_left_over;
+        const uint32_t _loopW = f_mgrW.is_left ? f_mgrW.frag_num - 1 : f_mgrW.frag_num;
 
         for (int k = 0; k < _loopW; ++k) {
             decx::vis::CPUK::_bilateral_rect_flex_uint8_ST(
@@ -442,7 +442,7 @@ void decx::vis::_bilateral_uint8_ST(const double* __restrict     src,
                 ker_dims, Wsrc, Wdst, reg_WL, _loop);
         }
         if (f_mgrW.is_left) {
-            const uint _sum_prev_lenW = proc_dim.x - f_mgrW.frag_left_over;
+            const uint32_t _sum_prev_lenW = proc_dim.x - f_mgrW.frag_left_over;
             decx::vis::CPUK::_bilateral_rect_flex_uint8_ST(
                 DECX_PTR_SHF_XY_SAME_TYPE(src, _sum_prev_lenH, _sum_prev_lenW * 2, Wsrc * 2),
                 _exp_chart_dist, _exp_chart_diff,
@@ -464,23 +464,23 @@ void decx::vis::_bilateral_uchar4_ST(const float* __restrict        src,
                                      float* __restrict              dst, 
                                      const uint2                    proc_dim, 
                                      const uint2                    ker_dims,
-                                     const uint                     Wsrc,
-                                     const uint                     Wdst,
+                                     const uint32_t                     Wsrc,
+                                     const uint32_t                     Wdst,
                                      const ushort                   reg_WL,
-                                     const uint                     _loop)
+                                     const uint32_t                     _loop)
 {
     __m256 res_vec8;
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     decx::utils::frag_manager f_mgrH, f_mgrW;
     decx::utils::frag_manager_gen_from_fragLen(&f_mgrH, proc_dim.y, _BLOCKED_CONV2_UINT8_H_);
     decx::utils::frag_manager_gen_from_fragLen(&f_mgrW, proc_dim.x, _BLOCKED_CONV2_UINT8_W_);
 
-    const uint _loopH = f_mgrH.is_left ? f_mgrH.frag_num - 1 : f_mgrH.frag_num;
+    const uint32_t _loopH = f_mgrH.is_left ? f_mgrH.frag_num - 1 : f_mgrH.frag_num;
     
     for (int i = 0; i < _loopH; ++i) 
     {
-        const uint _loopW = f_mgrW.is_left ? f_mgrW.frag_num - 1 : f_mgrW.frag_num;
+        const uint32_t _loopW = f_mgrW.is_left ? f_mgrW.frag_num - 1 : f_mgrW.frag_num;
 
         for (int k = 0; k < _loopW; ++k) {
             decx::vis::CPUK::_bilateral_rect_fixed_uchar4_ST(
@@ -490,7 +490,7 @@ void decx::vis::_bilateral_uchar4_ST(const float* __restrict        src,
                 ker_dims, Wsrc, Wdst, reg_WL, _loop);
         }
         if (f_mgrW.is_left) {
-            const uint _sum_prev_lenW = proc_dim.x - f_mgrW.frag_left_over;
+            const uint32_t _sum_prev_lenW = proc_dim.x - f_mgrW.frag_left_over;
             decx::vis::CPUK::_bilateral_rect_flex_uchar4_ST(
                 DECX_PTR_SHF_XY_SAME_TYPE(src, i * _BLOCKED_CONV2_UINT8_H_, _sum_prev_lenW * 4, Wsrc),
                 _exp_chart_dist, _exp_chart_diff,
@@ -502,8 +502,8 @@ void decx::vis::_bilateral_uchar4_ST(const float* __restrict        src,
     
     if (f_mgrH.is_left)
     {
-        const uint _sum_prev_lenH = proc_dim.y - f_mgrH.frag_left_over;
-        const uint _loopW = f_mgrW.is_left ? f_mgrW.frag_num - 1 : f_mgrW.frag_num;
+        const uint32_t _sum_prev_lenH = proc_dim.y - f_mgrH.frag_left_over;
+        const uint32_t _loopW = f_mgrW.is_left ? f_mgrW.frag_num - 1 : f_mgrW.frag_num;
 
         for (int k = 0; k < _loopW; ++k) {
             decx::vis::CPUK::_bilateral_rect_flex_uchar4_ST(
@@ -514,7 +514,7 @@ void decx::vis::_bilateral_uchar4_ST(const float* __restrict        src,
                 ker_dims, Wsrc, Wdst, reg_WL, _loop);
         }
         if (f_mgrW.is_left) {
-            const uint _sum_prev_lenW = proc_dim.x - f_mgrW.frag_left_over;
+            const uint32_t _sum_prev_lenW = proc_dim.x - f_mgrW.frag_left_over;
             decx::vis::CPUK::_bilateral_rect_flex_uchar4_ST(
                 DECX_PTR_SHF_XY_SAME_TYPE(src, _sum_prev_lenH, _sum_prev_lenW * 4, Wsrc),
                 _exp_chart_dist, _exp_chart_diff,
@@ -536,17 +536,17 @@ void decx::vis::_bilateral_uint8_caller(const double*               src,
                                         double*                     dst, 
                                         const uint2                 proc_dim, 
                                         const uint2                 neighbor_dims,
-                                        const uint                  Wsrc,
-                                        const uint                  Wdst,
+                                        const uint32_t                  Wsrc,
+                                        const uint32_t                  Wdst,
                                         const ushort                reg_WL,
                                         decx::utils::_thr_1D*       t1D,
                                         decx::utils::frag_manager*  f_mgr,
-                                        const uint                  _loop)
+                                        const uint32_t                  _loop)
 {
     const double* tmp_src_ptr = src;
     double *tmp_dst_ptr = dst;
-    size_t frag_src = (size_t)f_mgr->frag_len * (size_t)Wsrc * 2;
-    size_t frag_dst = (size_t)f_mgr->frag_len * (size_t)Wdst * 2;
+    uint64_t frag_src = (uint64_t)f_mgr->frag_len * (uint64_t)Wsrc * 2;
+    uint64_t frag_dst = (uint64_t)f_mgr->frag_len * (uint64_t)Wdst * 2;
 
     for (int i = 0; i < t1D->total_thread - 1; ++i) {
         t1D->_async_thread[i] = decx::cpu::register_task_default(decx::vis::_bilateral_uint8_ST,
@@ -574,17 +574,17 @@ void decx::vis::_bilateral_uchar4_caller(const float*               src,
                                         float*                      dst, 
                                         const uint2                 proc_dim, 
                                         const uint2                 neighbor_dims,
-                                        const uint                  Wsrc,
-                                        const uint                  Wdst,
+                                        const uint32_t                  Wsrc,
+                                        const uint32_t                  Wdst,
                                         const ushort                reg_WL,
                                         decx::utils::_thr_1D*       t1D,
                                         decx::utils::frag_manager*  f_mgr,
-                                        const uint                  _loop)
+                                        const uint32_t                  _loop)
 {
     const float* tmp_src_ptr = src;
     float *tmp_dst_ptr = dst;
-    size_t frag_src = (size_t)f_mgr->frag_len * (size_t)Wsrc;
-    size_t frag_dst = (size_t)f_mgr->frag_len * (size_t)Wdst;
+    uint64_t frag_src = (uint64_t)f_mgr->frag_len * (uint64_t)Wsrc;
+    uint64_t frag_dst = (uint64_t)f_mgr->frag_len * (uint64_t)Wdst;
 
     for (int i = 0; i < t1D->total_thread - 1; ++i) {
         t1D->_async_thread[i] = decx::cpu::register_task_default(decx::vis::_bilateral_uchar4_ST,
